Prediction/requirements.txt
scikit-learn>=1.1.3
keras>=2.10.0
tensorflow>=2.6.0
paho-mqtt